+++ backend/main.py
@@ -5,6 +5,7 @@ from auth import get_db
 
 import routers.assignments
 import routers.submissions
+import routers.grades
 import routers.courses
 import routers.materials
 import routers.parents
@@ -34,6 +35,7 @@ app = FastAPI(
 )
 app.include_router(routers.assignments.router)
 app.include_router(routers.submissions.router)
+app.include_router(routers.grades.router)
 app.include_router(routers.courses.router)
 app.include_router(routers.materials.router)
 app.include_router(routers.parents.router)
